cleanup bt nodes and move GoalsEmptyCondition out of archive
