+++ src/ros/rover/auto/auto_bringup/params/urc/nav2/bt_navigator.yaml
@@ -34,6 +34,7 @@ bt_navigator:
       - nova_get_current_pose_action_bt_node
       - nova_place_search_goals_action_bt_node
       - nova_object_detected_condition_bt_node
+      - nova_goals_empty_condition_bt_node
 
 bt_navigator_navigate_through_poses_rclcpp_node:
   ros__parameters:
