+++ src/ros/rover/auto/nova_behavior_tree/CMakeLists.txt
@@ -97,11 +97,11 @@ list(APPEND plugin_libs nova_get_dist_between_poses_action_bt_node)
 add_library(nova_ar_tag_detected_condition_bt_node SHARED plugins/condition/ar_tag_detected_condition.cpp)
 list(APPEND plugin_libs nova_ar_tag_detected_condition_bt_node)
 
-add_library(nova_enable_detection_condition_bt_node SHARED plugins/condition/enable_detection_condition.cpp)
-list(APPEND plugin_libs nova_enable_detection_condition_bt_node)
-
 add_library(nova_object_detected_condition_bt_node SHARED plugins/condition/object_detected_condition.cpp)
 list(APPEND plugin_libs nova_object_detected_condition_bt_node)
+
+add_library(nova_goals_empty_condition_bt_node SHARED plugins/condition/goals_empty_condition.cpp)
+list(APPEND plugin_libs nova_goals_empty_condition_bt_node)
 
 foreach(bt_plugin ${plugin_libs})
   ament_target_dependencies(${bt_plugin} ${dep_pkgs})
