+++ src/ros/rover/auto/nova_behavior_tree/include/nova_behavior_tree/archive/action/remove_passed_goals_and_cubes_action.hpp
