+++ src/ros/rover/auto/nova_behavior_tree/include/nova_behavior_tree/archive/condition/check_detected_condition.hpp
