+++ src/ros/rover/auto/nova_behavior_tree/include/nova_behavior_tree/archive/condition/compute_search_goals_condition.hpp
