+++ src/ros/rover/auto/nova_behavior_tree/include/nova_behavior_tree/archive/condition/cube_detected_condition.hpp
