+++ src/ros/rover/auto/nova_behavior_tree/include/nova_behavior_tree/archive/condition/emit_data_action_base.hpp
