+++ src/ros/rover/auto/nova_behavior_tree/include/nova_behavior_tree/archive/condition/enable_detection_condition.hpp
