+++ src/ros/rover/auto/nova_behavior_tree/include/nova_behavior_tree/archive/condition/first_goal_reached_condition.hpp
