+++ src/ros/rover/auto/nova_behavior_tree/include/nova_behavior_tree/archive/condition/get_final_goal_condition.hpp
