+++ src/ros/rover/auto/nova_behavior_tree/include/nova_behavior_tree/condition/goals_empty_condition.hpp
