+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/action/remove_passed_goals_and_ar_tags_action.cpp
