+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/action/remove_passed_goals_and_cubes_action.cpp
