+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/action/save_cube_poses_action.cpp
