+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/action/save_image_service.cpp
