+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/condition/check_detected_condition.cpp
