+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/condition/compute_search_goals_condition.cpp
