+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/condition/cube_detected_condition.cpp
