+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/condition/enable_detection_condition.cpp
