+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/condition/first_goal_reached_condition.cpp
