+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/condition/get_final_goal_condition.cpp
