+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/condition/get_first_goal_condition.cpp
