+++ src/ros/rover/auto/nova_behavior_tree/plugins/archive/condition/query_position_condition.cpp
